+++ apps/kyb-app/src/components/atoms/Stepper/components/atoms/Breadcrumbs/helpers/pick-wrapper.props.ts
@@ -11,6 +11,7 @@ export const pickWrapperProps: ElementPropsPicker<BreadcrumbsWrapperProps> = (
 
   const props: BreadcrumbsWrapperProps = {
     className: ctw(themeParams.className, { [themeParams.activeClassName || '']: active }),
+    style: active ? themeParams.activeStyles : themeParams.styles,
   };
 
   return props;