+++ apps/kyb-app/src/components/organisms/DynamicUI/TransitionListener/TransitionListener.tsx
@@ -1,6 +1,7 @@
-import { usePageResolverContext } from '@/components/organisms/DynamicUI/PageResolver/hooks/usePageResolverContext';
 import { useStateManagerContext } from '@/components/organisms/DynamicUI/StateManager/components/StateProvider';
 import { useUIElementToolsLogic } from '@/components/organisms/DynamicUI/hooks/useUIStateLogic/hooks/useUIElementsStateLogic/hooks/useUIElementToolsLogic';
+import { UIPage } from '@/domains/collection-flow';
+import { useRefValue } from '@/hooks/useRefValue';
 import { AnyChildren } from '@ballerine/ui';
 import { useEffect, useRef } from 'react';
 
@@ -11,37 +12,49 @@ export interface TransitionListenerTools {
 export interface TransitionListenerProps {
   onPrevious?: (tools: TransitionListenerTools, prevState: string, currentState: string) => void;
   onNext?: (tools: TransitionListenerTools, prevState: string, currentState: string) => void;
-  children: AnyChildren;
+  onFinish?: (tools: TransitionListenerTools, currentState: string) => void;
+  pages: UIPage[];
+  children: AnyChildren | ((tools: TransitionListenerTools) => AnyChildren);
 }
 
-export const TransitionListener = ({ onPrevious, onNext, children }: TransitionListenerProps) => {
+export const TransitionListener = ({
+  onPrevious,
+  onNext,
+  onFinish,
+  children,
+  pages,
+}: TransitionListenerProps) => {
   const { state } = useStateManagerContext();
-  const { pages } = usePageResolverContext();
   const { setElementCompleted } = useUIElementToolsLogic('');
 
   const prevStateRef = useRef(state);
-  const helpersRef = useRef({ setElementCompleted });
-
-  useEffect(() => {
-    helpersRef.current = {
-      setElementCompleted,
-    };
-  }, [setElementCompleted]);
+  const helpersRef = useRefValue(setElementCompleted);
 
   useEffect(() => {
     const currentPageIndex = pages.findIndex(page => page.stateName === state);
     const prevPageIndex = pages.findIndex(page => page.stateName === prevStateRef.current);
 
     if (currentPageIndex < prevPageIndex) {
-      onPrevious && onPrevious(helpersRef.current, prevStateRef.current, state);
+      onPrevious &&
+        onPrevious({ setElementCompleted: helpersRef.current }, prevStateRef.current, state);
     }
 
     if (currentPageIndex > prevPageIndex) {
-      onNext && onNext(helpersRef.current, prevStateRef.current, state);
+      onNext && onNext({ setElementCompleted: helpersRef.current }, prevStateRef.current, state);
     }
 
-    prevStateRef.current = state;
+    if (currentPageIndex !== -1) {
+      prevStateRef.current = state;
+    }
   }, [prevStateRef, helpersRef, state, pages]);
 
-  return <>{children}</>;
+  useEffect(() => {
+    const currentPageIndex = pages.findIndex(page => page.stateName === state);
+
+    if (currentPageIndex === -1 && state === 'finish') {
+      onFinish && onFinish({ setElementCompleted: helpersRef.current }, prevStateRef.current);
+    }
+  }, [state, pages, helpersRef, prevStateRef]);
+
+  return <>{typeof children === 'function' ? children({ setElementCompleted }) : children}</>;
 };