+++ packages/workflow-core/src/lib/plugins/external-plugin/api-plugin.ts
@@ -8,7 +8,7 @@ export class ApiPlugin {
   public static pluginKind = 'api';
   name: string;
   stateNames: string[];
-  url: string;
+  url: IApiPluginParams['url'];
   method: IApiPluginParams['method'];
   vendor?: IApiPluginParams['vendor'];
   headers: IApiPluginParams['headers'];
@@ -59,15 +59,15 @@ export class ApiPlugin {
         }
       }
 
-      const urlWithoutPlaceholders = await this.replaceValuePlaceholders(this.url, context);
+      const _url = await this._getPluginUrl(context);
 
       logger.log('API Plugin - Sending API request', {
-        url: urlWithoutPlaceholders,
+        url: _url,
         method: this.method,
       });
 
       const apiResponse = await this.makeApiRequest(
-        urlWithoutPlaceholders,
+        _url,
         this.method,
         requestPayload,
         await this.composeRequestHeaders(this.headers!, context),
@@ -75,7 +75,7 @@ export class ApiPlugin {
 
       logger.log('API Plugin - Received response', {
         status: apiResponse.statusText,
-        url: urlWithoutPlaceholders,
+        url: _url,
       });
 
       if (apiResponse.ok) {
@@ -115,6 +115,33 @@ export class ApiPlugin {
     }
   }
 
+  protected async _getPluginUrl(context: AnyRecord) {
+    let _url: string;
+
+    if (typeof this.url === 'string') {
+      _url = this.url;
+    } else {
+      // expected url to be an object { url: string, options: Record<string, string> }
+
+      if (this.url.url) {
+        _url = this.url.url;
+      } else {
+        throw new Error('URL is required');
+      }
+
+      const { options } = this.url;
+
+      if (options !== null && typeof options === 'object' && !Array.isArray(options)) {
+        _url = await this.replaceVariablesFromContext(this.url.url, this.url.options);
+      } else {
+        // if options is not an object
+        throw new Error('Url options should be an object');
+      }
+    }
+
+    return await this.replaceAllVariables(_url, context);
+  }
+
   returnSuccessResponse(callbackAction: string, responseBody: AnyRecord) {
     return { callbackAction, responseBody };
   }
@@ -142,11 +169,7 @@ export class ApiPlugin {
     };
 
     if (payload) {
-      for (const key of Object.keys(payload)) {
-        if (typeof payload[key] === 'string') {
-          payload[key] = await this.replaceValuePlaceholders(payload[key] as string, payload);
-        }
-      }
+      payload = await this._onPreparePayload(payload);
 
       // @TODO: Use an enum over string literals for HTTP methods
       if (method.toUpperCase() !== 'GET') {
@@ -170,6 +193,18 @@ export class ApiPlugin {
     }
 
     return res;
+  }
+
+  protected async _onPreparePayload(_payload: AnyRecord) {
+    const returnObj = JSON.parse(JSON.stringify(_payload));
+
+    for (const key of Object.keys(returnObj)) {
+      if (typeof returnObj[key] === 'string') {
+        returnObj[key] = await this.replaceAllVariables(returnObj[key] as string, returnObj);
+      }
+    }
+
+    return returnObj;
   }
 
   async transformData(transformers: Transformers | undefined, record: AnyRecord) {
@@ -218,14 +253,45 @@ export class ApiPlugin {
       Object.entries(headers).map(async header => [
         header[0],
         // eslint-disable-next-line @typescript-eslint/no-unsafe-argument
-        await this.replaceValuePlaceholders(header[1], context),
+        await this.replaceAllVariables(header[1], context),
       ]),
     );
 
     return Object.fromEntries(headersEntries);
   }
 
-  async replaceValuePlaceholders(content: string, context: TContext) {
+  async _onReplaceVariable(variableKey: string, replacedContent: string, placeholder: string) {
+    let replacedSecrets = await this.replaceCustomerVariable(
+      variableKey,
+      replacedContent,
+      placeholder,
+    );
+
+    // // TODO: Remove this line when migrate to new ballerine plugins
+    replacedSecrets = await this.replaceBallerineVariable(
+      variableKey,
+      replacedContent,
+      placeholder,
+    );
+
+    return replacedSecrets;
+  }
+
+  async replaceBallerineVariable(variableKey: string, content: string, placeholder: string) {
+    return await this.replaceSecretsByProvider('ballerine', variableKey, content, placeholder);
+  }
+
+  async replaceCustomerVariable(variableKey: string, content: string, placeholder: string) {
+    return await this.replaceSecretsByProvider('customer', variableKey, content, placeholder);
+  }
+
+  async replaceAllVariables(content: string, context: TContext) {
+    const replacedContent = await this.replaceSecrets(content);
+
+    return this.replaceVariablesFromContext(replacedContent, context);
+  }
+
+  async replaceSecrets(content: string) {
     const placeholders = content.match(/{(.*?)}/g);
 
     if (!placeholders) return content;
@@ -235,24 +301,58 @@ export class ApiPlugin {
     for (const placeholder of placeholders) {
       const variableKey = placeholder.replace(/{|}/g, '');
 
-      const isSystemSecret = variableKey.includes('secret.');
-      const isSecret = variableKey.includes('secrets.');
+      replacedContent = await this._onReplaceVariable(variableKey, replacedContent, placeholder);
+    }
 
-      if (isSystemSecret) {
-        const secretKey = variableKey.replace('secret.', '');
-        const secretValue = `${this.getSystemSecret(secretKey)}`;
+    return replacedContent;
+  }
 
-        replacedContent = replacedContent.replace(placeholder, secretValue);
-      } else if (isSecret) {
-        const secretKey = variableKey.replace('secrets.', '');
-        const secretValue = `${await this.fetchSecret(secretKey)}`;
+  async replaceVariablesFromContext(content: string, context: TContext) {
+    const placeholders = content.match(/{(.*?)}/g);
 
-        replacedContent = replacedContent.replace(placeholder, secretValue);
-      } else {
-        const placeholderValue = `${this.fetchObjectPlaceholderValue(context, variableKey)}`;
+    if (!placeholders) return content;
 
-        replacedContent = replacedContent.replace(placeholder, placeholderValue);
+    let replacedContent = content;
+
+    for (const placeholder of placeholders) {
+      const variableKey = placeholder.replace(/{|}/g, '');
+
+      if (variableKey.startsWith('secret')) {
+        continue;
       }
+
+      const placeholderValue = this.fetchObjectPlaceholderValue(context, variableKey);
+
+      if (placeholderValue === undefined) {
+        continue;
+      }
+
+      replacedContent = replacedContent.replace(placeholder, `${placeholderValue}`);
+    }
+
+    return replacedContent;
+  }
+
+  async replaceSecretsByProvider(
+    provider: 'ballerine' | 'customer',
+    variableKey: string,
+    content: string,
+    placeholder: string,
+  ) {
+    const variableName = provider === 'ballerine' ? 'secret.' : 'secrets.';
+
+    let replacedContent = content;
+
+    if (provider === 'ballerine' && variableKey.includes(variableName)) {
+      const secretKey = variableKey.replace(variableName, '');
+      const secretValue = `${this.getSystemSecret(secretKey)}`;
+
+      replacedContent = content.replace(placeholder, secretValue);
+    } else if (provider === 'customer' && variableKey.includes(variableName)) {
+      const secretKey = variableKey.replace('secrets.', '');
+      const secretValue = `${await this.fetchSecret(secretKey)}`;
+
+      replacedContent = content.replace(placeholder, secretValue);
     }
 
     return replacedContent;