+++ services/workflows-service/src/workflow/workflow.service.ts
@@ -112,20 +112,11 @@ import { StorageService } from '@/storage/storage.service';
 import { TOcrImages, UnifiedApiClient } from '@/common/utils/unified-api-client/unified-api-client';
 import { AwsSecretsManager } from '@/secrets-manager/aws-secrets-manager';
 import { InMemorySecretsManager } from '@/secrets-manager/in-memory-secrets-manager';
+import { FEATURE_LIST } from '@/customer/types';
 
 type TEntityId = string;
 
 export type TEntityType = 'endUser' | 'business';
-
-// TODO: TEMP (STUB)
-const policies = {
-  kycSignup: () => {
-    return [{ workflowDefinitionId: 'COLLECT_DOCS_b0002zpeid7bq9aaa', version: 1 }] as const;
-  },
-  kybSignup: () => {
-    return [{ workflowDefinitionId: 'COLLECT_DOCS_b0002zpeid7bq9bbb', version: 1 }] as const;
-  },
-};
 
 @Injectable()
 export class WorkflowService {
@@ -162,7 +153,6 @@ export class WorkflowService {
       version: true,
       definition: true,
       definitionType: true,
-      backend: true,
       extensions: true,
       persistStates: true,
       submitStates: true,
@@ -175,9 +165,7 @@ export class WorkflowService {
         contextSchema: data.contextSchema as InputJsonValue,
         documentsSchema: data.documentsSchema as InputJsonValue,
         config: data.config as InputJsonValue,
-        supportedPlatforms: data.supportedPlatforms as InputJsonValue,
         extensions: data.extensions as InputJsonValue,
-        backend: data.backend as InputJsonValue,
         persistStates: data.persistStates as InputJsonValue,
         submitStates: data.submitStates as InputJsonValue,
       },
@@ -202,9 +190,7 @@ export class WorkflowService {
       definition: true,
       contextSchema: true,
       config: true,
-      supportedPlatforms: true,
       extensions: true,
-      backend: true,
       persistStates: true,
       submitStates: true,
     };
@@ -714,7 +700,6 @@ export class WorkflowService {
       version: true,
       definition: true,
       definitionType: true,
-      backend: true,
       extensions: true,
       persistStates: true,
       submitStates: true,
@@ -2585,12 +2570,14 @@ export class WorkflowService {
     projectId: string;
     documentId: string;
   }) {
-    const ocrResult = await this.prismaService.$transaction(
+    return await this.prismaService.$transaction(
       async transaction => {
         const customer = await this.customerService.getByProjectId(projectId);
 
-        if (customer.features?.['isDocumentOcrEnabled'] === true) {
-          throw new BadRequestException('Document OCR is not enabled for this customer');
+        if (!customer.features?.[FEATURE_LIST.DOCUMENT_OCR]) {
+          throw new BadRequestException(
+            `Document OCR is not enabled for customer id ${customer.id}`,
+          );
         }
 
         const document = await this.findDocumentById({
@@ -2644,7 +2631,5 @@ export class WorkflowService {
         timeout: 180_000,
       },
     );
-
-    return ocrResult;
   }
 }