+++ packages/eslint-config/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @ballerine/eslint-config
 
+## 1.1.21
+
+### Patch Changes
+
+- bump
+
+## 1.1.20
+
+### Patch Changes
+
+- bump
+
+## 1.1.19
+
+### Patch Changes
+
+- version bump
+
 ## 1.1.18
 
 ### Patch Changes