+++ sdks/web-ui-sdk/CHANGELOG.md
@@ -1,5 +1,29 @@
 # web-ui-sdk
 
+## 1.5.40
+
+### Patch Changes
+
+- bump
+- Updated dependencies
+  - @ballerine/common@0.9.39
+
+## 1.5.39
+
+### Patch Changes
+
+- bump
+- Updated dependencies
+  - @ballerine/common@0.9.38
+
+## 1.5.38
+
+### Patch Changes
+
+- version bump
+- Updated dependencies
+  - @ballerine/common@0.9.37
+
 ## 1.5.37
 
 ### Patch Changes